fix: change style location
