+++ app/(tabs)/styles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    titleContainer: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        gap: 8,
+    },
+    stepContainer: {
+        gap: 8,
+        marginBottom: 8,
+    },
+    reactLogo: {
+        height: 178,
+        width: 290,
+        bottom: 0,
+        left: 0,
+        position: 'absolute',
+    },
+    block: {
+        marginTop: 32,
+    },
+});
